move present video feed logic into a utility
